app/requirements.txt
Flask==1.1.1
azureml-core
azureml.dataprep
azure-keyvault-secrets 
azure-keyvault-keys
azure-keyvault-certificates
azure-identity